Assignment_1_AI_Related/openrouter_flask_app/requirements.txt
flask
openai>=1.0.0
python-dotenv
